+++ languages/ruby/bitwarden_sdk/lib/bitwarden_error.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module BitwardenSDK
+  class BitwardenError < StandardError
+    def initialize(message = "SDK Error Occurred")
+      super(message)
+    end
+  end
+end
