+++ languages/ruby/bitwarden_sdk/lib/command_runner.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module BitwardenSDK
+  class CommandRunner
+    def initialize(bitwarden_sdk, handle)
+      @bitwarden_sdk = bitwarden_sdk
+      @handle = handle
+    end
+
+    # @param [Dry-Struct] cmd
+    def run(cmd)
+      @bitwarden_sdk.run_command(cmd.to_json, @handle)
+    end
+  end
+end
