+++ languages/ruby/bitwarden_sdk/lib/version.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module BitwardenSDK
+  VERSION = '0.1.2'
+end
+
