+++ languages/ruby/bitwarden_sdk/sig/bitwarden_client.rbs
@@ -1,3 +1,5 @@
+require_relative '../lib/schemas'
+
 class BitwardenClient
   @command_runner: CommandRunner
