+++ languages/ruby/bitwarden_sdk/sig/sdk.rbs
@@ -1,0 +1,3 @@
+module BitwardenSDK
+  VERSION: String
+end
