+++ languages/ruby/bitwarden_sdk/sig/secrets_client.rbs
@@ -1,4 +1,4 @@
-require_once '../extended_schemas/schemas.rbs'
+require_once '../lib/extended_schemas/schemas.rbs'
 require_once '../schemas.rbs'
 
 class SecretsClient
