--- languages/ruby/lib/bitwarden_error.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Bitwarden
-  module SDK
-    class BitwardenError < StandardError
-      def initialize(message = "SDK Error Occurred")
-        super(message)
-      end
-    end
-  end
-end
