--- languages/ruby/lib/ruby/sdk.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'sdk/version'
-
-  module Bitwarden
-    module SDK
-      class Error < StandardError; end
-      # Your code goes here...
-    end
-  end
