--- languages/ruby/lib/ruby/sdk/version.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Bitwarden
-  module SDK
-    VERSION = '0.1.10'
-  end
-end
-
