--- languages/ruby/sig/ruby/sdk.rbs
@@ -1,6 +1,0 @@
-module Bitwarden
-  module SDK
-    VERSION: String
-    # See the writing guide of rbs: https://github.com/ruby/rbs#guides
-  end
-end
